backend/pyproject.toml
[project]
name = "prizms-backend"
version = "0.1.0"
description = "Multi-round debate LLM tool with multi-provider support (Ollama, vLLM, LM Studio)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3",
    "langchain-core>=0.3",
    "langchain-openai>=0.2",
    "langchain-anthropic>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-community>=0.3",
    "langgraph>=0.2",
    "langchain-text-splitters>=0.3",
    "rich>=13.7.0",
    "pyyaml>=6.0",
    "pydantic-settings>=2.0.0",
    "supabase>=2.0.0",
    "pyjwt>=2.8.0",
    "email-validator>=2.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sse-starlette>=2.1.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
dev = [
    "httpx>=0.27.0",
]

[project.scripts]
prizms-api = "run_api:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "providers", "shared", "modules", "api"]
